feat: 위젯 전역 관리 스토어 추가 및 useHandleOpenWidget 훅으로 확장
