+++ src/shared/model/index.ts
@@ -1,6 +1,7 @@
 'use client';
 
-import { useHandleOpenChat } from './useHandleOpenChat';
+import { useHandleOpenWidget } from './useHandleOpenWidget';
+import { useWidgetStore } from './useWidgetStore';
 import { useThemeStore } from './useThemeStore';
 import { useMediaQuery } from './useMediaQuery';
 import { useIsMobile } from './useIsMobile';
@@ -9,7 +10,8 @@ import { useFABStore } from './useFABStore';
 export {
   useThemeStore,
   useFABStore,
-  useHandleOpenChat,
+  useHandleOpenWidget,
   useIsMobile,
   useMediaQuery,
+  useWidgetStore,
 };
