+++ src/shared/model/useWidgetStore.ts
@@ -1,0 +1,12 @@
+import { create } from 'zustand';
+
+type WidgetType = 'chat' | 'notification' | null;
+
+interface WidgetState {
+  activeWidget: WidgetType;
+  setActiveWidget: (widget: WidgetType) => void;
+}
+export const useWidgetStore = create<WidgetState>((set) => ({
+  activeWidget: null,
+  setActiveWidget: (widget) => set({ activeWidget: widget }),
+}));
